requirements.txt
openai>=1.0.0
streamlit
transformers
flask
requests
numpy
pandas
scikit-learn
matplotlib
seaborn
python-dotenv
transformers
torch
huggingface-hub
streamlit_feedback
replicate
gpt4all
langchain
PyPDF2
tiktoken
faiss-cpu
youtube-transcript-api
tqdm
pytube